+++ fw/nkymodem_cmd.c
@@ -47,9 +47,9 @@ const nk_direct_base_t ymodem_file_fpga =
     .area_base = FLASH_FPGA_BUFFER_MAIN,
     .erase_size = 65536, // Must be at least NK_MCUFLASH_ERASE_SIZE
     .info = NULL,
-    .flash_read = (int (*)(const void *, uint32_t, uint8_t *, uint32_t))nk_mcuflash_read, // info, address, data, byte_count
+    .flash_read = (int (*)(const void *, uint32_t, uint8_t *, size_t))nk_mcuflash_read, // info, address, data, byte_count
     .flash_erase = (int (*)(const void *info, uint32_t addr, uint32_t size))nk_mcuflash_erase,
-    .flash_write = (int (*)(const void *, uint32_t, const uint8_t *, uint32_t))nk_mcuflash_write, // info, address, data, byte_count
+    .flash_write = (int (*)(const void *, uint32_t, const uint8_t *, size_t))nk_mcuflash_write, // info, address, data, byte_count
     .granularity = 1
 };
 
@@ -59,9 +59,9 @@ const nk_direct_base_t ymodem_file_fw0 =
     .area_base = FLASH_FIRMWARE_BUFFER_0,
     .erase_size = 65536, // Must be at least NK_MCUFLASH_ERASE_SIZE
     .info = NULL,
-    .flash_read = (int (*)(const void *, uint32_t, uint8_t *, uint32_t))nk_mcuflash_read, // info, address, data, byte_count
+    .flash_read = (int (*)(const void *, uint32_t, uint8_t *, size_t))nk_mcuflash_read, // info, address, data, byte_count
     .flash_erase = (int (*)(const void *info, uint32_t addr, uint32_t size))nk_mcuflash_erase,
-    .flash_write = (int (*)(const void *, uint32_t, const uint8_t *, uint32_t))nk_mcuflash_write, // info, address, data, byte_count
+    .flash_write = (int (*)(const void *, uint32_t, const uint8_t *, size_t))nk_mcuflash_write, // info, address, data, byte_count
     .granularity = 1
 };
 
@@ -71,9 +71,9 @@ const nk_direct_base_t ymodem_file_fw1 =
     .area_base = FLASH_FIRMWARE_BUFFER_1,
     .erase_size = 65536, // Must be at least NK_MCUFLASH_ERASE_SIZE
     .info = NULL,
-    .flash_read = (int (*)(const void *, uint32_t, uint8_t *, uint32_t))nk_mcuflash_read, // info, address, data, byte_count
+    .flash_read = (int (*)(const void *, uint32_t, uint8_t *, size_t))nk_mcuflash_read, // info, address, data, byte_count
     .flash_erase = (int (*)(const void *info, uint32_t addr, uint32_t size))nk_mcuflash_erase,
-    .flash_write = (int (*)(const void *, uint32_t, const uint8_t *, uint32_t))nk_mcuflash_write, // info, address, data, byte_count
+    .flash_write = (int (*)(const void *, uint32_t, const uint8_t *, size_t))nk_mcuflash_write, // info, address, data, byte_count
     .granularity = 1
 };
